Make PENNANT 64-bit safe.

+++ src/ExportGold.cc
@@ -120,16 +120,16 @@ void ExportGold::writeGeoFile(
     } // if mype == 0
 
     // gather node info to PE 0
-    const int nump = mesh->nump;
+    const long long nump = mesh->nump;
     const double2* px = mesh->px;
 
-    int gnump = nump;
+    long long gnump = nump;
     Parallel::globalSum(gnump);
-    vector<int> penump(mype == 0 ? numpe : 0);
+    vector<long long> penump(mype == 0 ? numpe : 0);
     Parallel::gather(nump, &penump[0]);
-    vector<int> peoffset(mype == 0 ? numpe + 1 : 1);
+    vector<long long> peoffset(mype == 0 ? numpe + 1 : 1);
     partial_sum(penump.begin(), penump.end(), &peoffset[1]);
-    int offset;
+    long long offset;
     Parallel::scatter(&peoffset[0], offset);
     vector<double2> gpx(mype == 0 ? gnump : 0);
     Parallel::gatherv(&px[0], nump, &gpx[0], &penump[0]);
@@ -139,21 +139,21 @@ void ExportGold::writeGeoFile(
         ofs << "coordinates" << endl;
         ofs << setw(10) << gnump << endl;
         ofs << setprecision(5);
-        for (int p = 0; p < gnump; ++p)
+        for (long long p = 0; p < gnump; ++p)
             ofs << setw(12) << gpx[p].x << endl;
-        for (int p = 0; p < gnump; ++p)
+        for (long long p = 0; p < gnump; ++p)
             ofs << setw(12) << gpx[p].y << endl;
         // Ensight expects z-coordinates, so write 0 for those
-        for (int p = 0; p < gnump; ++p)
+        for (long long p = 0; p < gnump; ++p)
             ofs << setw(12) << 0. << endl;
     } // if mype
 
     const int* znump = mesh->znump;
-    const int* mapsp1 = mesh->mapsp1;
+    const long long* mapsp1 = mesh->mapsp1;
 
-    const int ntris = tris.size();
-    const int nquads = quads.size();
-    const int nothers = others.size();
+    const long long ntris = tris.size();
+    const long long nquads = quads.size();
+    const long long nothers = others.size();
 
     if (mype == 0) {
         pentris.resize(numpe);
@@ -168,20 +168,20 @@ void ExportGold::writeGeoFile(
     gnquads = accumulate(penquads.begin(), penquads.end(), 0);
     gnothers = accumulate(penothers.begin(), penothers.end(), 0);
 
-    vector<int> pesizes(mype == 0 ? numpe : 0);
+    vector<long long> pesizes(mype == 0 ? numpe : 0);
 
     // gather triangle info to PE 0
-    vector<int> trip(3 * ntris);
-    vector<int> gtris(gntris), gtrip(3 * gntris);
+    vector<long long> trip(3 * ntris);
+    vector<long long> gtris(gntris), gtrip(3 * gntris);
     Parallel::gatherv(&tris[0], ntris, &gtris[0], &pentris[0]);
     if (mype == 0) {
-        for (int pe = 0; pe < numpe; ++pe)
+        for (long long pe = 0; pe < numpe; ++pe)
             pesizes[pe] = pentris[pe] * 3;
     }
-    for (int t = 0; t < ntris; ++t) {
-        int z = tris[t];
-        int sbase = mapzs[z];
-        for (int i = 0; i < 3; ++i) {
+    for (long long t = 0; t < ntris; ++t) {
+        long long z = tris[t];
+        long long sbase = mapzs[z];
+        for (long long i = 0; i < 3; ++i) {
             trip[t * 3 + i] = mapsp1[sbase + i] + offset;
         }
     }
@@ -191,27 +191,27 @@ void ExportGold::writeGeoFile(
     if (mype == 0 && gntris > 0) {
         ofs << "tria3" << endl;
         ofs << setw(10) << gntris << endl;
-        for (int t = 0; t < gntris; ++t)
+        for (long long t = 0; t < gntris; ++t)
             ofs << setw(10) << gtris[t] + 1 << endl;
-        for (int t = 0; t < gntris; ++t) {
-            for (int i = 0; i < 3; ++i)
+        for (long long t = 0; t < gntris; ++t) {
+            for (long long i = 0; i < 3; ++i)
                 ofs << setw(10) << gtrip[t * 3 + i] + 1;
             ofs << endl;
         }
     } // if mype == 0 ...
 
     // gather quad info to PE 0
-    vector<int> quadp(4 * nquads);
-    vector<int> gquads(gnquads), gquadp(4 * gnquads);
+    vector<long long> quadp(4 * nquads);
+    vector<long long> gquads(gnquads), gquadp(4 * gnquads);
     Parallel::gatherv(&quads[0], nquads, &gquads[0], &penquads[0]);
     if (mype == 0) {
-        for (int pe = 0; pe < numpe; ++pe)
+        for (long long pe = 0; pe < numpe; ++pe)
             pesizes[pe] = penquads[pe] * 4;
     }
-    for (int q = 0; q < nquads; ++q) {
-        int z = quads[q];
-        int sbase = mapzs[z];
-        for (int i = 0; i < 4; ++i) {
+    for (long long q = 0; q < nquads; ++q) {
+        long long z = quads[q];
+        long long sbase = mapzs[z];
+        for (long long i = 0; i < 4; ++i) {
             quadp[q * 4 + i] = mapsp1[sbase + i] + offset;
         }
     }
@@ -221,45 +221,45 @@ void ExportGold::writeGeoFile(
     if (mype == 0 && gnquads > 0) {
         ofs << "quad4" << endl;
         ofs << setw(10) << gnquads << endl;
-        for (int q = 0; q < gnquads; ++q)
+        for (long long q = 0; q < gnquads; ++q)
             ofs << setw(10) << gquads[q] + 1 << endl;
-        for (int q = 0; q < gnquads; ++q) {
-            for (int i = 0; i < 4; ++i)
+        for (long long q = 0; q < gnquads; ++q) {
+            for (long long i = 0; i < 4; ++i)
                 ofs << setw(10) << gquadp[q * 4 + i] + 1;
             ofs << endl;
         }
     } // if mype == 0 ...
 
     // gather other info to PE 0
-    vector<int> othernump(nothers), otherp;
-    vector<int> gothers(gnothers), gothernump(gnothers);
+    vector<long long> othernump(nothers), otherp;
+    vector<long long> gothers(gnothers), gothernump(gnothers);
     Parallel::gatherv(&others[0], nothers, &gothers[0], &penothers[0]);
-    for (int n = 0; n < nothers; ++n) {
-        int z = others[n];
-        int sbase = mapzs[z];
+    for (long long n = 0; n < nothers; ++n) {
+        long long z = others[n];
+        long long sbase = mapzs[z];
         othernump[n] = znump[z];
-        for (int i = 0; i < znump[z]; ++i) {
+        for (long long i = 0; i < znump[z]; ++i) {
             otherp.push_back(mapsp1[sbase + i] + offset);
         }
     }
     Parallel::gatherv(&othernump[0], nothers, &gothernump[0], &penothers[0]);
-    int size = otherp.size();
+    long long size = otherp.size();
     Parallel::gather(size, &pesizes[0]);
-    int gsize = accumulate(pesizes.begin(), pesizes.end(), 0);
-    vector<int> gotherp(gsize);
+    long long gsize = accumulate(pesizes.begin(), pesizes.end(), 0);
+    vector<long long> gotherp(gsize);
     Parallel::gatherv(&otherp[0], size, &gotherp[0], &pesizes[0]);
 
     // write others
     if (mype == 0 && gnothers > 0) {
         ofs << "nsided" << endl;
         ofs << setw(10) << gnothers << endl;
-        for (int n = 0; n < gnothers; ++n)
+        for (long long n = 0; n < gnothers; ++n)
             ofs << setw(10) << gothers[n] + 1 << endl;
-        for (int n = 0; n < gnothers; ++n)
+        for (long long n = 0; n < gnothers; ++n)
             ofs << setw(10) << gothernump[n] << endl;
-        int gp = 0;
-        for (int n = 0; n < gnothers; ++n) {
-            for (int i = 0; i < gothernump[n]; ++i)
+        long long gp = 0;
+        for (long long n = 0; n < gnothers; ++n) {
+            for (long long i = 0; i < gothernump[n]; ++i)
                 ofs << setw(10) << gotherp[gp + i] + 1;
             ofs << endl;
             gp += gothernump[n];
@@ -297,13 +297,13 @@ void ExportGold::writeVarFile(
         ofs << setw(10) << 1 << endl;
     } // if mype == 0
 
-    int ntris = tris.size();
-    int nquads = quads.size();
-    int nothers = others.size();
+    long long ntris = tris.size();
+    long long nquads = quads.size();
+    long long nothers = others.size();
 
     // gather values on triangles to PE 0
     vector<double> tvar(ntris), gtvar(gntris);
-    for (int t = 0; t < ntris; ++t) {
+    for (long long t = 0; t < ntris; ++t) {
         tvar[t] = var[tris[t]];
     }
     Parallel::gatherv(&tvar[0], ntris, &gtvar[0], &pentris[0]);
@@ -311,14 +311,14 @@ void ExportGold::writeVarFile(
     // write values on triangles
     if (mype == 0 && gntris > 0) {
         ofs << "tria3" << endl;
-        for (int t = 0; t < gntris; ++t) {
+        for (long long t = 0; t < gntris; ++t) {
             ofs << setw(12) << gtvar[t] << endl;
         }
     } // if mype == 0 ...
 
     // gather values on quads to PE 0
     vector<double> qvar(nquads), gqvar(gnquads);
-    for (int q = 0; q < nquads; ++q) {
+    for (long long q = 0; q < nquads; ++q) {
         qvar[q] = var[quads[q]];
     }
     Parallel::gatherv(&qvar[0], nquads, &gqvar[0], &penquads[0]);
@@ -326,14 +326,14 @@ void ExportGold::writeVarFile(
     // write values on quads
     if (mype == 0 && gnquads > 0) {
         ofs << "quad4" << endl;
-        for (int q = 0; q < gnquads; ++q) {
+        for (long long q = 0; q < gnquads; ++q) {
             ofs << setw(12) << gqvar[q] << endl;
         }
     } // if mype == 0 ...
 
     // gather values on others to PE 0
     vector<double> ovar(nothers), govar(gnothers);
-    for (int n = 0; n < nothers; ++n) {
+    for (long long n = 0; n < nothers; ++n) {
         ovar[n] = var[others[n]];
     }
     Parallel::gatherv(&ovar[0], nothers, &govar[0], &penothers[0]);
@@ -341,7 +341,7 @@ void ExportGold::writeVarFile(
     // write values on others
     if (mype == 0 && gnothers > 0) {
         ofs << "nsided" << endl;
-        for (int n = 0; n < gnothers; ++n) {
+        for (long long n = 0; n < gnothers; ++n) {
             ofs << setw(12) << govar[n] << endl;
         }
     } // if mype == 0 ...
@@ -353,15 +353,15 @@ void ExportGold::writeVarFile(
 
 void ExportGold::sortZones() {
 
-    const int numz = mesh->numz;
+    const long long numz = mesh->numz;
     const int* znump = mesh->znump;
 
     mapzs.resize(numz);
 
     // sort zones by size, create an inverse map
-    int scount = 0;
-    for (int z = 0; z < numz; ++z) {
-        int zsize = znump[z];
+    long long scount = 0;
+    for (long long z = 0; z < numz; ++z) {
+        long long zsize = znump[z];
         if (zsize == 3)
             tris.push_back(z);
         else if (zsize == 4)
